telegram_bot/requirements.txt
python-telegram-bot>=22.6
aiosqlite>=0.22.1
python-dotenv>=1.2.2
redis>=7.2.1
aiohttp>=3.13.3
pytest>=9.0.2
pytest-asyncio>=1.3.0
loguru>=0.7.3
aiogram>=3.26.0
sqlalchemy>=2.1.0b1
aiofiles>=25.1.0
cryptography>=46.0.5
pydantic>=2.12.5
asyncpg>=0.31.0
psycopg2-binary>=2.9.11
flask>=3.1.3